+++ lib/flex_checkout_flutter.dart
@@ -3,3 +3,4 @@ library flex_checkout_flutter;
 export 'src/flex.dart';
 export 'src/types.dart';
 export 'src/flex_split_rent_button.dart';
+export 'src/flex_status_card.dart';